renamed preset_manager_simple
